Crate timer_list

A list of timed events that will be triggered sequentially when the timer expires.

Examples

use timer_list::{TimerEvent, TimerEventFn, TimerList};
use std::time::{Duration, Instant};

let mut timer_list = TimerList::new();

// set a timer that will be triggered after 1 second
let start_time = Instant::now();
timer_list.set(Duration::from_secs(1), TimerEventFn::new(|now| {
    println!("timer event after {:?}", now);
}));

while !timer_list.is_empty() {
    // check if there is any event that is expired
    let now = Instant::now().duration_since(start_time);
    if let Some((deadline, event)) = timer_list.expire_one(now) {
        // trigger the event, will print "timer event after 1.00s"
        event.callback(now);
        break;
    }
    std::thread::sleep(Duration::from_millis(10)); // relax the CPU
}

Structs

TimerEventFn

A simple wrapper of a closure that implements the TimerEvent trait.

TimerList

A list of timed events.

Traits

TimerEvent

A trait that all timed events must implement.

Type Aliases

TimeValue

The type of the time value.